RF: Move interface.run to core.local.run
Move run into the location proposed by gh-3192.

For now, keep the old module around for compatibility, particularly
with extensions.  Following the approach of a290c7b58, use the
warnings module rather than logging to warn about use of the obsolete
module.  (Note, though, that a DeprecationWarning is used, and
warnings in that category aren't shown by default.)

Closes #3434.
